% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks}
\alias{chunks}
\title{chunks function}
\usage{
chunks(x, range)
}
\arguments{
\item{x}{data.frame to be chunked}

\item{ranges}{vector with row numbers defining beginnings and ends of data chunks}
}
\description{
Extract a chunk of data from mgf data
What is returned: dataframe x is returned from ranges that are set
 in another dataframe: range
}
