% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractFeats}
\alias{extractFeats}
\title{Extract Feats function}
\usage{
extractFeats(df)
}
\arguments{
\item{df}{data.frame to be extracted}
}
\description{
Converts a data.frame to a list of features
}
